split project in api + example, add run paper for easy local development, add plugin.yml plugin
