+++ api/src/main/java/me/delta/mc/marker/api/Area.java
@@ -1,0 +1,11 @@
+package me.delta.mc.marker.api;
+
+public enum Area {
+  BOTTOM,
+  TOP,
+  NEGATIVE_X,
+  NEGATIVE_Z,
+  POSITIVE_X,
+  POSITIVE_Z,
+  FULL
+}
