+++ api/src/test/java/me/delta/mc/marker/MarkerExample.java
@@ -6,19 +6,19 @@ import me.delta.mc.marker.examples.GroundRGB;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.java.JavaPlugin;
 
-public final class MarkerAPI extends JavaPlugin {
+public class MarkerExample extends JavaPlugin {
 
     @Override
     public void onEnable() {
 
         Bukkit.getPluginManager().registerEvents(new AreaSelection(), this);
         Bukkit.getPluginManager().registerEvents(new GroundRGB(new RGBController(this, 1, 15)), this);
+
     }
 
     @Override
     public void onDisable() {
         // Plugin shutdown logic
     }
-
 
 }
