+++ settings.gradle.kts
@@ -1,5 +1,3 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
-
 rootProject.name = "MarkerAPI"
+
+include(":api", ":example")
